+++ Editor/JSONCImporter/JSONCImporter.cs
@@ -1,0 +1,19 @@
+#if UNITY_EDITOR
+using System.IO;
+using UnityEditor.AssetImporters;
+using UnityEngine;
+
+
+//this is a simple wrapper that allows us to
+//import "JSONC" files as text assets like json and txt files
+[ScriptedImporter(1, "jsonc")]
+public class JSONCImporter : ScriptedImporter
+{
+    public override void OnImportAsset(AssetImportContext ctx)
+    {
+        TextAsset subAsset = new TextAsset(File.ReadAllText(ctx.assetPath));
+        ctx.AddObjectToAsset("text", subAsset);
+        ctx.SetMainObject(subAsset);
+    }
+}
+#endif